move the get all edges and nodes to the task instead of client
