+++ tests/neptune_integration/test_neptune_client.py
@@ -31,84 +31,6 @@ class TestNeptuneClient(BaseNeptuneSessionClientTestCase):
         created_edge_name = get_test_graph(None).E().has(self.key_name, 'test_edge_1').values('name').next()
         self.assertEqual(created_edge_name, 'Name2')
 
-    def test_get_number_of_edges_grouped_by_label(self):
-        self.client.upsert_node('test_1', 'label', {})
-        self.client.upsert_node('test_2', 'label', {})
-        self.client.upsert_node('test_3', 'label', {})
-        self.client.upsert_edge('test_1', 'test_2', 'test_edge_1', 'edgy_edge', {})
-        self.client.upsert_edge('test_1', 'test_1', 'test_edge_1', 'self_reflection', {})
-        result_list = self.client.get_number_of_edges_grouped_by_label()
-        self.assertEqual(2, len(result_list))
-        edgy_edge_result = [result for result in result_list if result.get('type') == 'edgy_edge'][0]
-        self_reflection_result = [result for result in result_list if result.get('type') == 'self_reflection'][0]
-        self.assertEqual(edgy_edge_result.get('count'), 1)
-        self.assertEqual(self_reflection_result.get('count'), 1)
-
-    def test_get_number_of_edges_grouped_by_label_filtered(self):
-        self.client.upsert_node('test_1', 'label', {})
-        self.client.upsert_node('test_2', 'label', {})
-        self.client.upsert_node('test_3', 'label', {})
-        self.client.upsert_edge('test_1', 'test_2', 'test_edge_1', 'edgy_edge', {
-            'type:String(single)': 'user'
-        })
-        self.client.upsert_edge('test_1', 'test_1', 'test_edge_1', 'self_reflection', {
-            'type:String(single)': 'job'
-        })
-        filter_properties = [
-            ('type', 'job', traversal.eq)
-        ]
-        result_list = self.client.get_number_of_edges_grouped_by_label(filter_properties=filter_properties)
-        self.assertEqual(1, len(result_list))
-        edgy_edge_result = [result for result in result_list if result.get('type') == 'self_reflection'][0]
-        self.assertEqual(edgy_edge_result.get('count'), 1)
-
-    def test_get_number_of_nodes_grouped_by_label(self):
-        self.client.upsert_node('test_1', 'label', {})
-        self.client.upsert_node('test_2', 'label', {})
-        self.client.upsert_node('test_3', 'label1', {})
-        self.client.upsert_node('test_4', 'label2', {})
-        self.client.upsert_node('test_5', 'label2', {})
-        self.client.upsert_node('test_6', 'label2', {})
-        result_list = self.client.get_number_of_nodes_grouped_by_label()
-        self.assertEqual(len(result_list), 3)
-        label_result = [result for result in result_list if result.get('type') == 'label'][0]
-        label1_result = [result for result in result_list if result.get('type') == 'label1'][0]
-        label2_result = [result for result in result_list if result.get('type') == 'label2'][0]
-        self.assertEqual(label_result.get('count'), 2)
-        self.assertEqual(label1_result.get('count'), 1)
-        self.assertEqual(label2_result.get('count'), 3)
-
-    def test_get_number_of_nodes_grouped_by_label_filtered(self):
-        self.client.upsert_node('test_1', 'label', {
-            'type:String(single)': 'user'
-        })
-        self.client.upsert_node('test_2', 'label', {
-            'type:String(single)': 'job'
-        })
-        self.client.upsert_node('test_3', 'label1', {
-            'type:String(single)': 'job'
-        })
-        self.client.upsert_node('test_4', 'label2', {
-            'type:String(single)': 'job'
-        })
-        self.client.upsert_node('test_5', 'label2', {
-            'type:String(single)': 'job'
-        })
-        self.client.upsert_node('test_6', 'label2', {
-            'type:String(single)': 'user'
-        })
-        filter_properties = [
-            ('type', 'job', traversal.eq)
-        ]
-        result_list = self.client.get_number_of_nodes_grouped_by_label(filter_properties=filter_properties)
-        self.assertEqual(len(result_list), 3)
-        label_result = [result for result in result_list if result.get('type') == 'label'][0]
-        label1_result = [result for result in result_list if result.get('type') == 'label1'][0]
-        label2_result = [result for result in result_list if result.get('type') == 'label2'][0]
-        self.assertEqual(label_result.get('count'), 1)
-        self.assertEqual(label1_result.get('count'), 1)
-        self.assertEqual(label2_result.get('count'), 2)
-
     def test_delete_all_edges(self):
         self.client.upsert_node('test_1', 'label', {})
         self.client.upsert_node('test_2', 'label', {})
